server/requirements.txt
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.7
python-dotenv==0.19.2
redis==4.1.4
simplejson==3.17.6
six==1.16.0
Werkzeug==2.0.3
wrapt==1.13.3